+++ package.json
@@ -61,6 +61,7 @@
     "@sentry/webpack-plugin": "^2.5.0",
     "babel-runtime": "^6.26.0",
     "chroma-js": "^2.4.2",
+    "circular-dependency-plugin": "^5.2.2",
     "classnames": "^2.3.1",
     "d3-scale-chromatic": "^3.0.0",
     "dayjs": "^1.11.7",
