+++ src/card/CardReducer.ts
@@ -15,7 +15,7 @@ import {
   UPDATE_REPORT_DATABASE,
 } from './CardActions';
 import { TOGGLE_CARD_SETTINGS } from './CardActions';
-import { createUUID } from '../dashboard/DashboardThunks';
+import { createUUID } from '../utils/uuid';
 
 const update = (state, mutations) => Object.assign({}, state, mutations);
