+++ webpack.config.js
@@ -1,5 +1,6 @@
 const ReactRefreshWebpackPlugin = require('@pmmmwh/react-refresh-webpack-plugin');
 const { sentryWebpackPlugin } = require('@sentry/webpack-plugin');
+const CircularDependencyPlugin = require('circular-dependency-plugin');
 const rules = [
   {
     test: /\.(js|jsx|ts|tsx)$/,
@@ -60,7 +61,20 @@ module.exports = (env) => {
             project: 'neodash',
           }),
         ]
-      : [new ReactRefreshWebpackPlugin()],
+      : [
+          new ReactRefreshWebpackPlugin(),
+          new CircularDependencyPlugin({
+            // exclude detection of files based on a RegExp
+            exclude: /a\.js|node_modules/,
+            // add errors to webpack instead of warnings
+            failOnError: true,
+            // allow import cycles that include an asyncronous import,
+            // e.g. via import(/* webpackMode: "weak" */ './file.js')
+            allowAsyncCycles: false,
+            // set the current working directory for displaying module paths
+            cwd: process.cwd(),
+          }),
+        ],
     ignoreWarnings: [/Failed to parse source map/],
   };
 };
